Major changes to file structure, improvements in the UI
