+++ web/src/components/content-manager/module-creation/blocks/SortableBlock.tsx
@@ -1,0 +1,43 @@
+import { GripVertical } from 'lucide-react'
+import { useSortable } from '@dnd-kit/sortable'
+import { CSS } from '@dnd-kit/utilities'
+
+export function SortableBlock({ id, children }: { readonly id: string; readonly children: React.ReactNode }) {
+    const {
+        attributes,
+        listeners,
+        setNodeRef,
+        setActivatorNodeRef,
+        transform,
+        transition,
+        isDragging,
+    } = useSortable({ id })
+
+    return (
+        <div
+            ref={setNodeRef}
+            style={{
+                transform: CSS.Transform.toString(transform),
+                transition,
+                opacity: isDragging ? 0.4 : 1,
+                position: 'relative',
+                zIndex: isDragging ? 10 : undefined,
+            }}
+            className="flex gap-2"
+        >
+            <button
+                ref={setActivatorNodeRef}
+                type="button"
+                className={`cursor-grab active:cursor-grabbing flex-shrink-0 touch-none mt-2.5 transition-colors ${
+                    isDragging ? 'text-slate-400' : 'text-slate-300 hover:text-slate-400'
+                }`}
+                aria-label="Drag to reorder block"
+                {...attributes}
+                {...listeners}
+            >
+                <GripVertical className="w-4 h-4" />
+            </button>
+            <div className="flex-1 min-w-0">{children}</div>
+        </div>
+    )
+}
