+++ web/src/components/content-manager/module-creation/blocks/index.ts
@@ -1,0 +1,6 @@
+export { BlockWarning } from './BlockWarning'
+export { SortableBlock } from './SortableBlock'
+export { AddBlockMenu } from './AddBlockMenu'
+export { BlockDragPreview } from './BlockDragPreview'
+export type { BlockAccent, BlockAccentTheme } from './themes'
+export { BLOCK_THEMES } from './themes'
