+++ web/src/components/content-manager/module-creation/sections/SectionTitleEditor.tsx
@@ -1,0 +1,71 @@
+import { AlertCircle, Pencil } from 'lucide-react'
+import type { Section } from '../types'
+import type { SectionTheme } from './themes'
+
+export function SectionTitleEditor({ section, onUpdate, onStopEdit, titleMissing, editing, inputRef, sizerRef, syncWidth, onStartEdit, theme }: {
+    readonly section: Section
+    readonly onUpdate: (patch: Partial<Section>) => void
+    readonly onStopEdit: () => void
+    readonly titleMissing: boolean
+    readonly editing: boolean
+    readonly inputRef: React.RefObject<HTMLInputElement | null>
+    readonly sizerRef: React.RefObject<HTMLSpanElement | null>
+    readonly syncWidth: () => void
+    readonly onStartEdit: () => void
+    readonly theme: SectionTheme
+}) {
+    if (editing) {
+        return (
+            <span className="relative z-10 inline-flex items-center min-w-0 shrink-0 max-w-[40%]">
+                <span
+                    ref={sizerRef}
+                    aria-hidden
+                    className="absolute invisible whitespace-pre text-sm font-semibold px-2 py-0.5 pointer-events-none"
+                    style={{ minWidth: '10ch' }}
+                />
+                <input
+                    ref={inputRef}
+                    value={section.title}
+                    onChange={e => { onUpdate({ title: e.target.value }); syncWidth() }}
+                    onBlur={onStopEdit}
+                    onKeyDown={e => { if (e.key === 'Enter' || e.key === 'Escape') { e.currentTarget.blur() } }}
+                    onClick={e => e.stopPropagation()}
+                    onDoubleClick={e => e.stopPropagation()}
+                    placeholder="Section title..."
+                    style={{ minWidth: '10ch' }}
+                    className={`text-sm font-semibold text-slate-800 bg-white rounded-md px-2 py-0.5 focus:outline-none w-full ${
+                        titleMissing
+                            ? 'border border-amber-400 focus:ring-amber-300/40'
+                            : `border ${theme.inputBorder} ${theme.inputRing}`
+                    }`}
+                />
+                <Pencil className={`w-3 h-3 ml-2 text-slate-300 ${theme.pencil} transition-colors flex-shrink-0`} />
+            </span>
+        )
+    }
+
+    return (
+        <button
+            type="button"
+            onClick={e => { e.stopPropagation(); onStartEdit() }}
+            onDoubleClick={e => e.stopPropagation()}
+            className="inline-flex items-center gap-1.5 min-w-0 group text-left relative z-10 shrink-0 max-w-[40%]"
+            title={titleMissing ? 'Title is required — click to add' : 'Click to edit title'}
+        >
+            {(() => {
+                let textCls = 'text-slate-400'
+                if (titleMissing) textCls = 'text-amber-500'
+                else if (section.title) textCls = 'text-slate-800'
+                return (
+                    <span className={`text-sm font-semibold truncate ${textCls}`}>
+                        {section.title || (titleMissing ? 'Title required' : 'Untitled section')}
+                    </span>
+                )
+            })()}
+            {titleMissing
+                ? <AlertCircle className="w-3 h-3 text-amber-400 flex-shrink-0" />
+                : <Pencil className={`w-3 h-3 text-slate-300 ${theme.pencil} transition-colors flex-shrink-0`} />
+            }
+        </button>
+    )
+}
