+++ web/src/components/content-manager/module-creation/sections/index.ts
@@ -1,0 +1,7 @@
+export { SectionCard } from './SectionCard'
+export { SectionDragPreview } from './SectionDragPreview'
+export { SectionRuleIcons } from './SectionRuleIcons'
+export { SectionTitleEditor } from './SectionTitleEditor'
+export { ViewTabToggle } from './ViewTabToggle'
+export type { AccentColor, SectionTheme } from './themes'
+export { THEMES } from './themes'
