+++ web/src/css/loading.css
@@ -1,0 +1,53 @@
+@keyframes breathe {
+  0%,
+  100% {
+    opacity: 0.4;
+    transform: scale(1.4);
+  }
+  50% {
+    opacity: 1;
+    transform: scale(1.65);
+  }
+}
+
+/* ── Full-page loading screen (used on initial mount of a route) ── */
+.loading-screen {
+  position: fixed;
+  inset: 0;
+  z-index: 50;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background-color: white;
+  gap: 2rem;
+  opacity: 0;
+  transition: opacity 0.8s ease;
+  pointer-events: none;
+}
+
+.loading-logo {
+  animation: breathe 2.8s ease-in-out infinite;
+}
+
+.loading-message {
+  font-size: 0.875rem;
+  letter-spacing: 0.1em;
+  color: #9ca3af;
+  text-align: center;
+  padding: 0 1rem;
+}
+
+/* ── Navigation transition overlay (lives at root, covers everything) ── */
+.navigation-loading-overlay {
+  position: fixed;
+  inset: 0;
+  z-index: 100;
+  background-color: white;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  opacity: 0;
+  transition: opacity 0.3s ease;
+  pointer-events: none;
+}
