+++ web/src/routes/__root.tsx
@@ -2,20 +2,22 @@ import { Navbar } from "@/components/navbar";
 import { createRootRoute, Outlet } from "@tanstack/react-router";
 import { Sidebar } from "@/components/sidebar";
 import ComplianceFlow from "@/components/compliance";
+import { NavigationLoadingProvider } from "@/lib/navigation-loading";
 
 const RootLayout = () => {
   return (
-    <div className="h-screen bg-white shadow-md rounded-xl">
-      <Navbar />
-      <div className="h-[92%] flex flex-row">
-        <Sidebar />
-        <div className="flex-1 overflow-y-auto">
-          <Outlet />
+    <NavigationLoadingProvider>
+      <div className="h-screen bg-white shadow-md rounded-xl">
+        <Navbar />
+        <div className="h-[92%] flex flex-row">
+          <Sidebar />
+          <div className="flex-1 overflow-y-auto">
+            <Outlet />
+          </div>
+          <ComplianceFlow />
         </div>
-        <ComplianceFlow />
       </div>
-      {/* <TanStackRouterDevtools /> */}
-    </div>
+    </NavigationLoadingProvider>
   );
 };
