+++ web/src/routes/statistics.tsx
@@ -13,16 +13,9 @@ export const Route = createFileRoute("/statistics")({
 
 function RouteComponent() {
   return (
-    <div className="h-full w-full overflow-y-auto bg-gradient-to-br from-slate-50 via-white to-purple-50/30 animate-[fadeIn_0.5s_ease-out]">
-      <style>{`
-        @keyframes fadeIn {
-          from { opacity: 0; transform: translateY(10px); }
-          to { opacity: 1; transform: translateY(0); }
-        }
-      `}</style>
-
+    <div className="h-full w-full overflow-y-auto bg-gray-50/50">
       {/* Header */}
-      <div className="h-[8%] w-full border-b border-slate-200/60 flex items-center px-6 backdrop-blur-sm bg-white/40">
+      <div className="h-[8%] w-full border-b border-gray-200 flex items-center px-6 bg-white">
         <StatsHeader />
       </div>
